5/3

+++ fnl_CapstoneProject/src/Resources.java
@@ -3,9 +3,9 @@ import java.awt.*;
 public interface Resources {
 	
 	public int returnPoints();
-	public double getX();
-	public double getY();
-	public double createPath();
+	public int getX();
+	public int getY();
+	public int createPath();
 	public Image returnPic();
 
 }
